requirements.txt
flask==3.1.3
flask-cors==6.0.2
opencv-python==4.10.1.26
deepface==0.0.98
tensorflow==2.18.0
numpy==1.26.4
pandas==2.2.0
scikit-learn==1.5.0
pillow==10.4.0
requests==2.32.3
python-dotenv==1.0.1
tqdm==4.66.4
matplotlib==3.9.0
scipy==1.13.0
protobuf==5.27.2